LayoutTests/accessibility/mac/live-regions/live-region-dynamically-added.html

<!DOCTYPE HTML><!-- webkit-test-runner [ IsAriaLiveRegionManagementEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/accessibility-helper.js"></script>
<meta http-equiv="content-language" content="en">
</head>
<body id="body">

<script>
var output = "This tests live regions that are added to the page dynamically behave correctly.\n\n";    
var notificationCount = 0;
var testFinished = false;

function notificationCallback(object, notification, userInfo) {
    if (notification == "AXAnnouncementRequested") {
        output += `Received announcement request:\n`;
        output += formatAnnouncementUserInfo(userInfo);
        output += `\n`;

        notificationCount++; 
        testFinished = true;
    }
}

if (accessibilityController) {
    window.jsTestIsAsync = true;
    accessibilityController.addNotificationListener(notificationCallback);
    accessibilityController.accessibleElementById("polite-region");

    // Add a live region, this shouldn't make any announcement.
    const newRegion = document.createElement("div");
    newRegion.setAttribute("aria-live", "polite");
    newRegion.setAttribute("id", "region");
    newRegion.textContent = "Alerts will appear here.";
    document.getElementById("body").appendChild(newRegion);
 
    setTimeout(async function() {
        await waitFor(() => accessibilityController.accessibleElementById("region") != null);
        
        // Now that the live region has been added to the page, any changes should be announced.
        newRegion.textContent = "Alert! Region changed";
        
        await waitFor(() => testFinished);
        accessibilityController.removeNotificationListener();
        debug(output);
        finishJSTest();
    }, 0);
}
</script>
</body>
</html>
