LayoutTests/accessibility/mac/live-regions/live-region-dynamically-added-expected.txt

This tests live regions that are added to the page dynamically behave correctly.

Received announcement request:
AnnouncementKey: Alert! Region changed{
    AXLanguage = en;
}
AXPriorityKey: 10
AXAnnouncementIsLiveRegionKey: true


PASS successfullyParsed is true

TEST COMPLETE
Alert! Region changed
